feat: add resources
