+++ simaple/fetch/inference/builtin/angelicbuster.yaml
@@ -1,0 +1,13 @@
+jobtype: angelicbuster
+passive:
+  attack_power: 91
+  magic_attack: 0
+  DEX: 100
+  damage_multiplier: 20
+candidates:
+  - - final_damage_multiplier: 25
+    - final_damage_multiplier: 26
+  - - null
+    - final_damage_multiplier: 18
+    - final_damage_multiplier: 12
+    - final_damage_multiplier: 6
