+++ simaple/fetch/inference/builtin/ark.yaml
@@ -1,0 +1,8 @@
+jobtype: ark
+passive:
+  attack_power: 112
+  magic_attack: 0
+  final_damage_multiplier: 39.755
+  STR: 60
+candidates:
+  - - null
