+++ simaple/fetch/inference/builtin/blaster.yaml
@@ -1,0 +1,21 @@
+jobtype: blaster
+passive:
+  attack_power: 62
+  magic_attack: 0
+  final_damage_multiplier: 5
+  attack_power_multiplier: 15
+  STR: 30
+  DEX: 30
+  damage_multiplier: 20
+candidates:
+  - - null
+    - final_damage_multiplier: 70
+    - final_damage_multiplier: 63
+    - final_damage_multiplier: 56
+    - final_damage_multiplier: 49
+    - final_damage_multiplier: 42
+    - final_damage_multiplier: 35
+    - final_damage_multiplier: 28
+    - final_damage_multiplier: 21
+    - final_damage_multiplier: 14
+    - final_damage_multiplier: 7
