+++ simaple/fetch/inference/builtin/cadena.yaml
@@ -1,0 +1,11 @@
+jobtype: cadena
+passive:
+  attack_power: 72
+  LUK: 80
+  DEX: 30
+candidates:
+  - - null # 오라웨폰
+    - final_damage_multiplier: 80
+    - final_damage_multiplier: 160
+  - - null
+    - final_damage_multiplier: 36
