+++ simaple/fetch/inference/builtin/darkknight.yaml
@@ -1,0 +1,11 @@
+jobtype: darkknight
+passive:
+  attack_power: 100 # 110 + (31)
+  magic_attack: 0
+  final_damage_multiplier: 164.384
+  STR: 50
+  DEX: 30
+  damage_multiplier: 5
+candidates:
+  - - null
+    - final_damage_multiplier: 6
