+++ simaple/fetch/inference/builtin/demonslayer.yaml
@@ -1,0 +1,11 @@
+jobtype: demonslayer
+passive:
+  attack_power: 101 # 110 + (31)
+  magic_attack: 0
+  final_damage_multiplier: 43
+  STR: 50
+  DEX: 30
+  damage_multiplier: 36
+candidates:
+  - - null
+    - final_damage_multiplier: 15
