+++ simaple/fetch/inference/builtin/eunwol.yaml
@@ -1,0 +1,11 @@
+jobtype: eunwol
+passive:
+  attack_power: 20
+  STR: 60
+  damage_multiplier: 20
+  final_damage_multiplier: 15
+candidates:
+  - - final_damage_multiplier: 16
+    - final_damage_multiplier: 17
+  - - null
+    - final_damage_multiplier: 20
