+++ simaple/fetch/inference/builtin/flamewizard.yaml
@@ -1,0 +1,10 @@
+jobtype: flamewizard
+passive:
+  magic_attack: 52
+  magic_attack_multiplier: 10
+  INT: 60
+  DEX: 30
+  final_damage_multiplier: 25
+candidates:
+  - - final_damage_multiplier: 50
+    - final_damage_multiplier: 53
