+++ simaple/fetch/inference/builtin/kain.yaml
@@ -1,0 +1,11 @@
+jobtype: kain
+passive:
+  attack_power: 184
+  DEX: 60
+  damage_multiplier: 14
+  final_damage_multiplier: 24
+candidates:
+  - - final_damage_multiplier: 21
+    - final_damage_multiplier: 20
+  - - final_damage_multiplier: 31
+    - final_damage_multiplier: 30
