+++ simaple/fetch/inference/builtin/kinesis.yaml
@@ -1,0 +1,11 @@
+jobtype: kinesis
+passive:
+  magic_attack: 104
+  magic_attack_multiplier: 10
+  INT: 40
+  DEX: 50
+  damage_multiplier: 22
+  final_damage_multiplier: 40
+candidates:
+  - - final_damage_multiplier: 38
+    - final_damage_multiplier: 41.52
