+++ simaple/fetch/inference/builtin/mechanic.yaml
@@ -1,0 +1,11 @@
+jobtype: mechanic
+passive:
+  attack_power: 76 # 110 + (31)
+  magic_attack: 0
+  final_damage_multiplier: 39
+  STR: 30
+  DEX: 30
+  damage_multiplier: 55
+candidates:
+  - - null
+    - final_damage_multiplier: 11
