+++ simaple/fetch/inference/builtin/mercedes.yaml
@@ -1,0 +1,21 @@
+jobtype: mercedes
+passive:
+  attack_power: 172 # 110 + (31)
+  magic_attack: 0
+  final_damage_multiplier: 66.375
+  attack_power_multiplier: 31
+  DEX: 50
+  STR: 30
+  damage_multiplier: 50
+candidates:
+  - - null
+    - final_damage_multiplier: 20
+    - final_damage_multiplier: 18
+    - final_damage_multiplier: 16
+    - final_damage_multiplier: 14
+    - final_damage_multiplier: 12
+    - final_damage_multiplier: 10
+    - final_damage_multiplier: 8
+    - final_damage_multiplier: 6
+    - final_damage_multiplier: 4
+    - final_damage_multiplier: 2
