+++ simaple/fetch/inference/builtin/nightlord.yaml
@@ -1,0 +1,14 @@
+jobtype: nightlord
+passive:
+  attack_power: 110
+  magic_attack: 0
+  final_damage_multiplier: 35
+  attack_power_multiplier: 10
+  LUK: 70
+  DEX: 30
+  damage_multiplier: 15
+candidates:
+  - - null
+    - final_damage_multiplier: 36
+  - - null
+    - final_damage_multiplier: 14
