+++ simaple/fetch/inference/builtin/nightwalker.yaml
@@ -1,0 +1,23 @@
+jobtype: nightwalker
+passive:
+  attack_power: 101 # 110 + (31)
+  magic_attack: 0
+  final_damage_multiplier: 29.6
+  LUK: 80
+  damage_multiplier: 30
+candidates:
+  - - null
+    - final_damage_multiplier: 36
+  - - null
+    - final_damage_multiplier: 20
+  - - null
+    - final_damage_multiplier: 20
+    - final_damage_multiplier: 18
+    - final_damage_multiplier: 16
+    - final_damage_multiplier: 14
+    - final_damage_multiplier: 12
+    - final_damage_multiplier: 10
+    - final_damage_multiplier: 8
+    - final_damage_multiplier: 6
+    - final_damage_multiplier: 4
+    - final_damage_multiplier: 2
