+++ simaple/fetch/inference/builtin/pathfinder.yaml
@@ -1,0 +1,11 @@
+jobtype: pathfinder
+passive:
+  attack_power: 82
+  attack_power_multiplier: 21
+  DEX: 132
+  STR: 30
+  damage_multiplier: 10
+  final_damage_multiplier: 20
+candidates:
+  - - null
+    - final_damage_multiplier: 8
