+++ simaple/fetch/inference/builtin/phantom.yaml
@@ -1,0 +1,14 @@
+jobtype: phantom
+passive:
+  attack_power: 81
+  LUK: 140
+  DEX: 30
+  damage_multiplier: 31
+  final_damage_multiplier: 71.6
+candidates:
+  - - null
+    - final_damage_multiplier: 36
+  - - null
+    - final_damage_multiplier: 20
+  - - null
+    - final_damage_multiplier: 40
