+++ simaple/fetch/inference/builtin/sniper.yaml
@@ -1,0 +1,11 @@
+jobtype: sniper
+passive:
+  attack_power: 70
+  magic_attack: 0
+  final_damage_multiplier: 66.87296
+  STR: 30
+  DEX: 132
+  damage_multiplier: 15
+candidates:
+  - - null
+    - final_damage_multiplier: 4
