+++ simaple/fetch/inference/builtin/striker.yaml
@@ -1,0 +1,15 @@
+jobtype: striker
+passive:
+  attack_power: 81
+  magic_attack: 0
+  final_damage_multiplier: 20.75
+  attack_power_multiplier: 10
+  STR: 80
+  damage_multiplier: 5
+candidates:
+  - - null
+    - final_damage_multiplier: 18
+    - final_damage_multiplier: 12
+    - final_damage_multiplier: 6
+  - - null
+    - final_damage_multiplier: 27
