+++ simaple/fetch/inference/builtin/viper.yaml
@@ -1,0 +1,11 @@
+jobtype: buccaneer
+passive:
+  attack_power: 92 # 110 + (31)
+  magic_attack: 0
+  final_damage_multiplier: 24.3
+  attack_power_multiplier: 16
+  STR: 50
+  DEX: 30
+candidates:
+  - - null
+    - final_damage_multiplier: 26
