+++ simaple/fetch/inference/builtin/wildhunter.yaml
@@ -1,0 +1,11 @@
+jobtype: wildhunter
+passive:
+  attack_power: 111
+  magic_attack: 0
+  attack_power_multiplier: 35
+  final_damage_multiplier: 32
+  DEX: 152
+  STR: 92
+  damage_multiplier: 10
+candidates:
+  - - null
