+++ tests/fetch/response/dumped/archmagefp.json
