+++ tests/fetch/response/dumped/archmagefp2.json
