+++ tests/fetch/response/dumped/luminous.json
