+++ tests/fetch/response/test_logic.py
@@ -4,14 +4,19 @@ import pytest
 from loguru import logger
 
 from simaple.core import BaseStatType, JobType
+from simaple.fetch.inference.attack_logic import predicate_attack_factor
 from simaple.fetch.inference.builtin_settings import get_predefined_setting
-from simaple.fetch.inference.logic import predicate_attack
+from simaple.fetch.inference.logic import infer_stat
 from simaple.fetch.inference.stat_logic import predicate_stat, predicate_stat_factor
 from simaple.fetch.response.character import CharacterResponse
 from simaple.fetch.translator.gear import GearTranslator
 from simaple.fetch.translator.kms.gear import kms_gear_stat_translator
 from simaple.fetch.translator.kms.potential import kms_potential_translator
 from simaple.gear.gear_repository import GearRepository
+
+
+def _get_file_path(file_name: str) -> str:
+    return f"tests/fetch/response/dumped/{file_name}"
 
 
 def _get_response(file_name: str) -> CharacterResponse:
@@ -30,14 +35,14 @@ def _get_response(file_name: str) -> CharacterResponse:
 
 @pytest.fixture(name="new_data")
 def fixture_new_data():
-    return _get_response("tests/fetch/response/fp.json")
+    return _get_response(_get_file_path("archmagefp2.json"))
 
 
 def test_attack_predication(new_data):
     logger.disable("")
 
     setting = get_predefined_setting(JobType.archmagefb)
-    result = predicate_attack(new_data, setting)
+    result = predicate_attack_factor(new_data, setting)
 
     assert [v for v, score in result][0] == (96.0, 126.0, 3316)
 
@@ -46,8 +51,8 @@ def test_luminous():
     logger.disable("")
 
     setting = get_predefined_setting(JobType.luminous)
-    result = predicate_attack(
-        _get_response("tests/fetch/response/backend.json"), setting
+    result = predicate_attack_factor(
+        _get_response(_get_file_path("luminous.json")), setting
     )
 
     assert result[0][0] == (77.0, 18.0, 827)
@@ -57,8 +62,8 @@ def test_infinity_triggered_predication():
     logger.disable("")
 
     setting = get_predefined_setting(JobType.archmagefb)
-    results = predicate_attack(
-        _get_response("tests/fetch/response/biomolecule.json"), setting
+    results = predicate_attack_factor(
+        _get_response(_get_file_path("archmagefp.json")), setting
     )
 
     assert results[0][0] == (96.0, 126.0, 3366)
@@ -67,8 +72,9 @@ def test_infinity_triggered_predication():
 @pytest.mark.parametrize(
     "filename",
     [
-        "tests/fetch/response/dumped/ilium.json",
-        "tests/fetch/response/dumped/bishop.json",
+        _get_file_path("ilium.json"),
+        _get_file_path("bishop.json"),
+        _get_file_path("windbreaker.json"),
     ],
 )
 def test_from_dumped(filename):
@@ -77,7 +83,7 @@ def test_from_dumped(filename):
     response = _get_response(filename)
     setting = get_predefined_setting(response.get_jobtype())
 
-    results = predicate_attack(response, setting)
+    results = predicate_attack_factor(response, setting)
     assert len(results) > 0
 
 
@@ -86,7 +92,10 @@ def test_stat():
 
     setting = get_predefined_setting(JobType.archmagefb)
     results = predicate_stat_factor(
-        _get_response("tests/fetch/response/fp.json"), setting, 270, BaseStatType.INT
+        _get_response(_get_file_path("archmagefp2.json")),
+        setting,
+        270,
+        BaseStatType.INT,
     )
 
     assert results[0][0] == (5902, 560, 20290)
@@ -97,7 +106,7 @@ def test_full_stat_predication():
 
     setting = get_predefined_setting(JobType.archmagefb)
     results = predicate_stat(
-        _get_response("tests/fetch/response/fp.json"),
+        _get_response(_get_file_path("archmagefp2.json")),
         setting,
         270,
         size=1,
@@ -119,3 +128,46 @@ def test_full_stat_predication():
         "INT_static": 20290.0,
         "DEX_static": 420.0,
     }
+
+
+def test_full_stat_predication2():
+    logger.disable("")
+
+    setting = get_predefined_setting(JobType.windbreaker)
+    results = predicate_stat(
+        _get_response(_get_file_path("windbreaker.json")),
+        setting,
+        277,
+        size=1,
+    )
+
+    best_result = results[0][0]
+
+    assert best_result.short_dict() == {
+        "STR": 2295.0,
+        "LUK": 1251.0,
+        "INT": 1138.0,
+        "DEX": 5326.0,
+        "STR_multiplier": 144.0,
+        "LUK_multiplier": 135.0,
+        "INT_multiplier": 159.0,
+        "DEX_multiplier": 519.0,
+        "STR_static": 733.0,
+        "LUK_static": 386.0,
+        "INT_static": 630.0,
+        "DEX_static": 20210.0,
+    }
+
+
+def test_infer_stat():
+    logger.disable("")
+
+    setting = get_predefined_setting(JobType.windbreaker)
+    results = infer_stat(
+        _get_response(_get_file_path("windbreaker.json")),
+        setting,
+        authentic_force=270,
+        size=5,
+    )
+
+    assert len(results) == 5
